♻️ Express Pagefind build as operations
